==> 3_MLR.Rmd <==
---
title: 'Multivariate analysis: multiple logistic regression'
author: Núria Jolis Orriols
date: '`r format(Sys.Date(),"%e de %B, %Y")`' 
output:
  pdf_document:
    toc: yes
    df_print: kable
header-includes:
  - \usepackage[english]{babel}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment = NULL, cache=TRUE, error = TRUE)
options(width=90)
```

```{r packages, message=FALSE, echo=FALSE, warning=FALSE}
libraries <- c("kernlab", "ggplot2" ,"caret", "tidyverse", "dplyr", "gmodels", "MASS")
check.libraries <- is.element(libraries, installed.packages()[, 1])==FALSE
libraries.to.install <- libraries[check.libraries]
if (length(libraries.to.install!=0)) {
  install.packages(libraries.to.install)
}

success <- sapply(libraries,require, quietly = FALSE,  character.only = TRUE)
if(length(success) != length(libraries)) {stop("A package failed to return a success in require() function.")}
```

# Step 1 - Obtaining datasets

```{r}
dataA <- read.csv("data/datasetA.csv",stringsAsFactors=T)
dataB <- read.csv("data/datasetB.csv",stringsAsFactors=T)
dataC <- read.csv("data/datasetC.csv",stringsAsFactors=T)
dataD <- read.csv("data/datasetD.csv",stringsAsFactors=T)
```

# Step 2 - Generate multiple logistic regresion models 

For each dataset the full model is first generated. After, the stepAIC() function of the MASS package is used to select the most contributing variables according to the lowest AIC.

## Data A

### Full logistic regression model

```{r}
xnames <- names(dataA[2:48])
foriA<- as.formula(paste("outcome ~ ", paste(xnames, collapse= "+")))
```


```{r}
full_mlrA<- glm(foriA, data = dataA, family = 'binomial')
summary(full_mlrA)
```
The full model with dataset A found the variables age, presence of deficiencyanemia, presence of renal.failure, temperature, platelets and PCO2 to be statistically significant.

### Perform stepwise variable selection

```{r, message=FALSE, warning=FALSE}
step_mlrA<- full_mlrA %>% stepAIC(trace = FALSE)
```

It is not possible to reduce the model with the stepAIC() function on the dataset A because there are missing values.

## Data B

### Full logistic regression model

```{r}
xnamesB <- names(dataB[2:48])
foriB<- as.formula(paste("outcome ~ ", paste(xnamesB, collapse= "+")))
```


```{r}
full_mlrB<- glm(foriB, data = dataB, family = 'binomial')
summary(full_mlrB)
```

The full model found the variables age, deficiencyanemia, renal.failure, temperature, platelets and PCO2 to be statistically significant.

### Perform stepwise variable selection

The stepAIC function of the MASS package will be used to select the most contributive variables:

```{r, message=FALSE, warning=FALSE}
step_mlrB <- full_mlrB %>% stepAIC(trace = FALSE)
summary(step_mlrB)
```

The function chose a final model with the following formula: 

```{r}
formula(step_mlrB)
cbind("Estimate"=coef(summary(step_mlrB))[,1], "Odds"=exp(coef(step_mlrB)), coef(summary(step_mlrB))[,2:4])%>% kable(format="pandoc", round(3))
```

## Data C

### Full logistic regression model

```{r}
xnamesC <- names(dataC[2:40])
foriC<- as.formula(paste("outcome ~ ", paste(xnamesC, collapse= "+")))
```


```{r}
full_mlrC<- glm(foriC, data = dataC, family = 'binomial')
summary(full_mlrC)
```

### Perform stepwise variable selection

The stepAIC function of the MASS package will be used to select the most contributive variables:

```{r, message=FALSE,warning=FALSE}
step_mlrC <- full_mlrC %>% stepAIC(trace = FALSE)
summary(step_mlrC)
```

```{r}
formula(step_mlrC)
cbind("Estimate"=coef(summary(step_mlrC))[,1], "Odds"=exp(coef(step_mlrC)), coef(summary(step_mlrC))[,2:4])%>% kable(format="pandoc", round(3))
```

## Data D

### Full logistic regression model

```{r, message=FALSE}
xnamesD <- names(dataD[2:48])
foriD<- as.formula(paste("outcome ~ ", paste(xnamesD, collapse= "+")))
```


```{r}
full_mlrD<- glm(foriD, data = dataD, family = 'binomial')
summary(full_mlrD)
```

The full model found the variables age, deficiencyanemia, renal.failure, temperature, platelets and PCO2 to be statistically significant.

### Perform stepwise variable selection

The stepAIC function of the MASS package will be used to select the most contributive variables:

```{r, message=FALSE,warning=FALSE}
step_mlrD <- full_mlrD %>% stepAIC(trace = FALSE)
summary(step_mlrD)
```

The function chose a final model with the following formula: 

```{r}
formula(step_mlrD)
cbind("Estimate"=coef(summary(step_mlrD))[,1], "Odds"=exp(coef(step_mlrD)), coef(summary(step_mlrD))[,2:4])%>% kable(format="pandoc", round(3))
```
